qrforge/requirements.txt
# QRForge Dependencies
qrcode[pil]>=7.4.2
pyzbar>=0.1.9
pillow>=10.0.0
